+++ internal/params/optional.go
@@ -23,7 +23,7 @@ func (b *optionalBuilder) EndOptional() Builder {
 	b.opt.parent.params = append(b.opt.parent.params, &Parameter{
 		parent: b.opt.parent,
 		name:   b.opt.name,
-		value:  value.OptionalValue(b.opt.value),
+		value:  b.opt.value,
 	})
 
 	return b.opt.parent